Add model comparison functionality and update README
- Introduced a new script `compare_all_models.py` for comparing four medical CLIP models simultaneously, including automated SHAP analysis and visualizations.
- Updated the README to reflect the new comparison option, detailing usage instructions and features of the new functionality.
- Added a reference to the new comparison module in the package's `__init__.py` file for better accessibility.
