+++ src/mmshap_medclip/__init__.py
@@ -2,5 +2,8 @@
 from . import models      # registra modelos
 from . import datasets    # registra datasets
 
+# Módulo de comparación de modelos
+from . import comparison  # funciones para comparar múltiples modelos
+
 # (opcional) expón utilidades si quieres:
 # from .registry import build_model, build_dataset
